Preventing the record with the repetitive username on the client side

+++ client/src/storage.rs
@@ -1,6 +1,5 @@
 use serde::{Serialize, Deserialize};
 use std::fs;
-use std::path::Path;
 
 #[derive(Serialize, Deserialize)]
 
@@ -11,10 +10,16 @@ pub struct LocalUserRecord {
     pub enc_x: Vec<u8>,
 }
 
+//maybe will modify to return Result<> instead of crashing with panic
+//might also modify to store base64 encoding instead of the bytes format
+
 pub fn store_local_user(record: &LocalUserRecord) {
-    let dir = "../data";
+    let dir = "client/data";
     let path = format!("{}/{}.json", dir, record.username);
     fs::create_dir_all(dir).expect("Failed to create the data directory");
+    if std::path::Path::new(&path).exists() {
+        panic!("Username already exists");
+    }
     let json = serde_json::to_string_pretty(record).expect("Failed to serialize the record");
     fs::write(path, json).expect("Failed to write the user record");
 }
